requirements.txt
telethon
python-dotenv
selenium
schedule